pyproject.toml
[tool.poetry]
name = "whisperx"
version = "3.1.1"
description = ""
authors = ["Max Bain"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
nltk = "^3.8.1"
transformers = {extras = ["torch"], version = "^4.36.2"}
pyannote-audio = {version = "^3.1.1"}
torch = "^2.1.2"
torchaudio = "^2.1.2"
faster-whisper = "^0.10.0"
dacite = "^1.8.1"

[tool.poetry.group.test.dependencies]
jiwer = "^3.0.3"
pytest = "^7.4.4"
datasets = "^2.16.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
isort = "*"
pyright = "^1"
mypy = "^1"
ruff = "*"

[tool.isort]
profile = "black"

[tool.ruff]
select = ['E', 'F', 'W', 'I', 'N', 'UP', 'E', 'W']
ignore = ['N815', 'N802']

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
typeCheckingMode = "strict"
useLibraryCodeForTypes = true

reportMissingImports = true
reportMissingTypeStubs = false

strictListInference = true
stringDictionaryInference = true
strictSetInference = true
reportUnusedImport = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
